+++ addon/components/light-tree-branch.js
@@ -1,5 +1,5 @@
 import Component from '@ember/component';
-import layout from '../templates/components/x-tree-branch';
+import layout from '../templates/components/light-tree-branch';
 
 export default Component.extend({
   layout,
