+++ addon/components/light-tree-children.js
@@ -1,5 +1,5 @@
 import Component from '@ember/component';
-import layout from '../templates/components/x-tree-children';
+import layout from '../templates/components/light-tree-children';
 
 export default Component.extend({
   layout,
