+++ tests/integration/components/x-tree-children-test.js
@@ -1,7 +1,7 @@
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
-moduleForComponent('x-tree-children', 'Integration | Component | x tree children', {
+moduleForComponent('light-tree-children', 'Integration | Component | light tree children', {
   integration: true
 });
 
@@ -9,15 +9,15 @@ test('it renders', function(assert) {
   // Set any properties with this.set('myProperty', 'value');
   // Handle any actions with this.on('myAction', function(val) { ... });
 
-  this.render(hbs`{{x-tree-children}}`);
+  this.render(hbs`{{light-tree-children}}`);
 
   assert.equal(this.$().text().trim(), '');
 
   // Template block usage:
   this.render(hbs`
-    {{#x-tree-children}}
+    {{#light-tree-children}}
       template block text
-    {{/x-tree-children}}
+    {{/light-tree-children}}
   `);
 
   assert.equal(this.$().text().trim(), 'template block text');
