+++ tests/integration/components/x-tree-node-test.js
@@ -1,7 +1,7 @@
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
-moduleForComponent('x-tree-node', 'Integration | Component | x tree node', {
+moduleForComponent('light-tree-node', 'Integration | Component | light tree node', {
   integration: true
 });
 
@@ -9,15 +9,15 @@ test('it renders', function(assert) {
   // Set any properties with this.set('myProperty', 'value');
   // Handle any actions with this.on('myAction', function(val) { ... });
 
-  this.render(hbs`{{x-tree-node}}`);
+  this.render(hbs`{{light-tree-node}}`);
 
   assert.equal(this.$().text().trim(), '');
 
   // Template block usage:
   this.render(hbs`
-    {{#x-tree-node}}
+    {{#light-tree-node}}
       template block text
-    {{/x-tree-node}}
+    {{/light-tree-node}}
   `);
 
   assert.equal(this.$().text().trim(), 'template block text');
