+++ tests/integration/components/x-tree-test.js
@@ -1,7 +1,7 @@
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
-moduleForComponent('x-tree', 'Integration | Component | x tree', {
+moduleForComponent('light-tree', 'Integration | Component | light tree', {
   integration: true
 });
 
@@ -9,15 +9,15 @@ test('it renders', function(assert) {
   // Set any properties with this.set('myProperty', 'value');
   // Handle any actions with this.on('myAction', function(val) { ... });
 
-  this.render(hbs`{{x-tree}}`);
+  this.render(hbs`{{light-tree}}`);
 
   assert.equal(this.$().text().trim(), '');
 
   // Template block usage:
   this.render(hbs`
-    {{#x-tree}}
+    {{#light-tree}}
       template block text
-    {{/x-tree}}
+    {{/light-tree}}
   `);
 
   assert.equal(this.$().text().trim(), 'template block text');
